requirement.txt
bcrypt==4.1.2
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.2.2
Flask-SQLAlchemy==3.0.5
itsdangerous==2.1.0
Flask-Migrate==4.0.5
pytz==2023.4
